20160520
大改Buff类的架构（都不知道第几次大改了……）。
小推几个技能。

+++ jlq_MBE_BattleSimulation/BuffGainDefence.cs
@@ -15,7 +15,7 @@ namespace JLQ_MBE_BattleSimulation
         /// <param name="time">持续时间</param>
         /// <param name="defenceGain">增益的防御值，负数则为降低</param>
         /// <param name="game">游戏对象</param>
-        public BuffGainDefence(Character buffee, Character buffer, int time, float defenceGain, Game game)
+        public BuffGainDefence(Character buffee, Character buffer, int time, double defenceGain, Game game)
             : base(buffee, buffer, time, "Defence", defenceGain, "坚固：防御", "破碎：防御", game)
         {
